kpathsea: fix warnings

+++ kpathsea/xgetcwd.c
@@ -118,7 +118,7 @@ xgetcwd (void)
         }
         if (!found)
             FATAL2("No inode %d/device %d in parent directory",
-                   cwd_stat.st_ino, cwd_stat.st_dev);
+                   (int) cwd_stat.st_ino, (int) cwd_stat.st_dev);
 
         xclosedir(parent_dir);
     }
